refactor interactors with inheritance
