+++ lib/core/interactor/action_plan_reader.rb
@@ -1,31 +1,13 @@
 module Core
-  class ActionPlanReader
-    attr_accessor :id
-    private :id=
-
-    def initialize(id)
-      self.id = id
-    end
-
-    def call
-      data        = Registry::Repository[:action_plan].find(id)
-      action_plan = ActionPlan.new(id, data)
-
-      if action_plan.valid?
-        action_plan.tap do |a|
-          a.categories = build_categories(a.categories)
-        end
-      elsif block_given?
-        yield
-      end
-    end
-
+  class ActionPlanReader < BaseContentReader
     private
 
-    def build_categories(category_ids)
-      category_ids.map do |category_id|
-        CategoryReader.new(category_id).call
-      end
+    def entity_class
+      ActionPlan
+    end
+
+    def repository
+      Registry::Repository[:action_plan]
     end
   end
 end
