+++ lib/core/interactor/corporate_reader.rb
@@ -1,19 +1,10 @@
 module Core
   class CorporateReader < ArticleReader
-    def call
-      data    = repository.find(id)
-      article = CorporateArticle.new(id, data)
-
-      if article.valid?
-        article.tap do |a|
-          a.categories = build_categories(a.categories)
-        end
-      elsif block_given?
-        yield
-      end
-    end
-
     private
+
+    def entity_class
+      CorporateArticle
+    end
 
     def repository
       Registry::Repository[:corporate]
